Moved the previous component called CaptionTextNodeList to a new component called CaptionTextNodeListContainer. Inside the previous component, CaptionTextNodeList, only render the list of CaptionTextNodes.
The purpose of this is to remove the logic that calculates the speaker name at the top of the list. This way we can add a shouldComponentUpdate hook to CaptionTextNodeList and prevent unnecessary re-renders on such a large list.
